requirements.txt
backoff==2.2.1
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
Deprecated==1.2.13
Flask==2.2.3
google-auth==2.17.2
googleapis-common-protos==1.59.0
grpcio==1.53.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kubernetes==26.1.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
networkx==3.1
oauthlib==3.2.2
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-exporter-zipkin==1.17.0
opentelemetry-exporter-zipkin-json==1.17.0
opentelemetry-exporter-zipkin-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-instrumentation-requests==0.38b0
opentelemetry-instrumentation-wsgi==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
